chore(squash-prep): rebase ui-loop-issues onto main + fix #555 (v0.5.635)
Replays the PR's diff onto main's current metadata so GitHub squash-merge
applies cleanly without the v0.5.608-637 patch-version collisions that
accumulated while the PR was in review.

- Cargo.toml workspace version: 0.5.634 (main HEAD) -> 0.5.635
- CLAUDE.md "Current Version": 0.5.634 -> 0.5.635
- CLAUDE.md "Recent Changes": single new v0.5.635 entry summarizing the PR
  (per CLAUDE.md's external-contributor policy: "the maintainer bumps the
  version and writes the changelog entry at merge time")
- CHANGELOG.md: single new v0.5.635 entry
- crates/perry-ui-macos/src/widgets/table.rs::get_filter_text: closes #555
  - return *const u8 (pinned StringHeader) instead of f64
  - drop the per-fn extern "C" decl with mismatched js_string_from_bytes len
  - mirror the canonical textfield::get_string_value pattern
- crates/perry-ui-macos/src/lib.rs::perry_ui_table_get_filter_text:
  signature flips from -> f64 to -> i64 to match perry-dispatch's
  ReturnKind::Str protocol (codegen reads x0, not d0)

Pre-existing bug #556 (Table renderCell with string-concat closure body
produces NSException) noted in CLAUDE.md but deferred — git blame shows
view_for_column was last touched 2026-02-28, three months before this PR.

+++ crates/perry-ui-macos/src/lib.rs
@@ -2167,8 +2167,8 @@ pub extern "C" fn perry_ui_table_set_filter_text(handle: i64, text_ptr: i64) {
 }
 
 #[no_mangle]
-pub extern "C" fn perry_ui_table_get_filter_text(handle: i64) -> f64 {
-    widgets::table::get_filter_text(handle)
+pub extern "C" fn perry_ui_table_get_filter_text(handle: i64) -> i64 {
+    widgets::table::get_filter_text(handle) as i64
 }
 
 // =============================================================================
